clippy update

+++ examples/MiMC7.rs
@@ -79,7 +79,7 @@ fn mimc7_circuit<F: FieldExt>(
 
                 let xkc_value = x_value + k_value + c_value;
                 ctx.assign(xkc, xkc_value);
-                ctx.assign(y, xkc_value.pow_vartime([7 as u64]));
+                ctx.assign(y, xkc_value.pow_vartime([7_u64]));
             });
         });
 
@@ -106,7 +106,7 @@ fn mimc7_circuit<F: FieldExt>(
 
                 let xkc_value = x_value + k_value + c_value;
                 ctx.assign(xkc, xkc_value);
-                ctx.assign(y, xkc_value.pow_vartime([7 as u64]));
+                ctx.assign(y, xkc_value.pow_vartime([7_u64]));
             });
         });
 
@@ -144,7 +144,7 @@ fn mimc7_circuit<F: FieldExt>(
                 // step 1 through 90: calculate witness values from iteration results
                 row_value += F::from(1);
                 x_value += k_value + c_value;
-                x_value = x_value.pow_vartime([7 as u64]);
+                x_value = x_value.pow_vartime([7_u64]);
                 c_value = F::from_str_vartime(round_key).unwrap();
                 // Step 1 through 90: assign witness values
                 ctx.add(&mimc7_step, (x_value, k_value, c_value, row_value));
@@ -153,7 +153,7 @@ fn mimc7_circuit<F: FieldExt>(
             // step 90: calculate final output
             row_value += F::from(1);
             x_value += k_value + c_value;
-            x_value = x_value.pow_vartime([7 as u64]);
+            x_value = x_value.pow_vartime([7_u64]);
             // Step 90: output the hash result as x + k in witness generation
             // output is not displayed as a public column, which will be implemented in the future
             ctx.add(&mimc7_last_step, (x_value, k_value, c_value, row_value)); // c_value is not
@@ -164,7 +164,6 @@ fn mimc7_circuit<F: FieldExt>(
     });
 
     let compiler = Compiler::new(SingleRowCellManager {}, SimpleStepSelectorBuilder {});
-    
     compiler.compile(&mimc7)
 
 }
